libast: cleanup: remove obsolete exec(3) fallbacks
execlp, execve and execvp have all been in POSIX since Issue 1, so
the need for those fallbacks expired many years ago.

execvpe is a GNU extension, so that fallback is not obsolete --
however, it is unused (even in the full original AST code base), so
let's remove that as well.