package: install: rename existing binaries to .old
This is needed to satisfy System Integrity Protection on macOS. It
will SIGKILL a binary upon invocation if it overwrote a previous
binary that was in use, unless the binary was renamed first.

Also, go back to the 'cp -PR' workaround for macOS /bin/cp to copy
a symlink instead of just 'cp -R', because POSIX says it is
unspecified whether -P, -H or -L is the default behaviour if -R is
specified, and we need to be sure to have -P to copy a symlink.