features/standards: use standards macros for BSDs (re: 058e5a84)
The features/standards iffe script claims:
/*
 * On BSD systems, _POSIX_SOURCE and such are used to *limit*
 * functionality to a known API; they don't enable anything. The
 * general intent in BSD is to enable everything by default.
 */
But this is completely false. I don't remember where on earth I
ever got that idea. All sorts of stuff is disabled by default and
needs to be enabled using _POSIX_SOURCE, _XOPEN_SORUCE, and (on
OpenBSD) _BSD_SOURCE or (on NetBSD) _NETBSD_SOURCE.

I discovered this when parallel building became impossible on all
the BSDs as of the referenced commit, because the SA_RESTART macro
is not defined by default on any of them. <sigh>