libast: features/fs: fix for NetBSD
There was an "implicit function definition" warning for getfsstat()
in port/mnt.c on NetBSD. On that OS, the function is gone from the
headers but still in the libraries, so iffe detects it so
port/mnt.c will use it. NetBSD has getvfsstat() instead, which is
the same function under a different name, just to be different.

src/lib/libast/features/fs:
- Add test for getvfsstat() modelled on the existing getfsstat()
  test with struct statvfs. If it is detected, the emitted code
  will define getfsstat as getvfsstat and the corresponding MNT_*
  macros as ST_* for getvfsstat, so that the port/mnt.c code will
  compile on NetBSD without further modification/complication.
- Chain the get(v)fsstat tests in an if...elif...elif...endif so
  that only one of them gets to emit results.